PPE_detection_YOLO/requirements_final.txt
Flask==3.0.0
opencv-python-headless==4.13.0.92
numpy>=1.26.0
Pillow>=10.0.0
python-dotenv==1.0.0
Werkzeug==3.0.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
